Add files via upload
Game is completed! I was rushing for time so there are some features I would have liked to add that I did not have time to, and there are some little hacks in the code that aren't very pretty (a.k.a a bit of hardcoding) but that get the job done as I did not have time to do some things with pointers I would have liked to do and I did not have time in the end to create more static variables.